\name{admissions}
\alias{admissions}
\title{Total Non-elective G&A Admissions (FFCEs)}
\description{UK NHS Total Non-elective G&A Admissions (FFCEs). April 2008 -- June 2013.}
\usage{admissions}
\format{Time series data}
\source{UK Department of Health: Monthly Activity Return}

\keyword{datasets}
\examples{
tsdisplay(admissions)
}