% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-utils.R
\name{check_variable_info}
\alias{check_variable_info}
\title{Check Variable Information}
\usage{
check_variable_info(variable_info)
}
\arguments{
\item{variable_info}{A data frame. The data frame should have at least the columns
"ensembl", "symbol", "uniprot", and "entrezid". Each of these columns should have
at least one non-NA value.}
}
\value{
This function does not return any value. It stops execution and throws an
error if any of the required conditions are not met.
}
\description{
This function checks if the input data frame \code{variable_info} has the required columns,
and if these columns have any non-NA values.
}
\examples{
variable_info <- data.frame(
  ensembl = c("ENSG000001", NA, NA),
  symbol = c("Gene1", "Gene2", "Gene3"),
  uniprot = c(NA, "P12345", "Q67890"),
  entrezid = c(101, 102, 103)
)
check_variable_info(variable_info)
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
