% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_llm_module_utils.R
\name{clear_output_dir}
\alias{clear_output_dir}
\title{Clear the embedding_output directory}
\usage{
clear_output_dir(output_dir = NULL)
}
\arguments{
\item{output_dir}{A character string specifying the directory to be checked and cleared.}
}
\value{
This function does not return a value. It prints a success or failure message
to the console based on whether the directory was found and cleared.
}
\description{
This internal function checks for the existence of a specified directory.
If the directory exists, it recursively deletes all files and subdirectories within it,
and then removes the empty directory itself. If the directory does not exist, it provides
a message indicating so.
}
\note{
This function is designed for internal use to manage temporary or output files.
Ensure that the specified directory does not contain critical data before clearing.
}
\author{
Feifan Zhang \email{FEIFAN004@e.ntu.edu.sg}
}
\keyword{internal}
