% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data_in_package.R
\docType{data}
\name{demo_mo_M_data}
\alias{demo_mo_M_data}
\title{Multi-Omics Demo Dataset: Metabolomics}
\format{
A data frame with KEGG compound identifiers (human/hsa) and
associated statistical measures from differential abundance analysis.
}
\usage{
demo_mo_M_data
}
\description{
Significantly altered metabolites from human metabolomics data, provided
as a demo for multi-omics pathway enrichment analysis.
Contains KEGG compound IDs with statistical measures.
}
\examples{
data(demo_mo_M_data)
head(demo_mo_M_data)

}
\keyword{datasets}
