% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-enrich_pathway.R
\name{enrich_pathway}
\alias{enrich_pathway}
\alias{enrich_pathway.mass_dataset}
\alias{enrich_pathway.data.frame}
\title{Perform pathway enrichment analysis}
\usage{
enrich_pathway(
  object,
  database = c("go", "kegg", "reactome"),
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)

\method{enrich_pathway}{mass_dataset}(
  object,
  database = c("go", "kegg", "reactome"),
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)

\method{enrich_pathway}{data.frame}(
  object,
  database = c("go", "kegg", "reactome"),
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)
}
\arguments{
\item{object}{A data frame or mass_dataset object containing gene identifiers.}

\item{database}{A vector of databases to use for enrichment (default: c("GO", "kegg", "reactome")).}

\item{path}{The directory where results should be saved (default: "result").}

\item{OrgDb}{The OrgDb object to use for GO enrichment.}

\item{organism}{The organism code for KEGG and Reactome enrichment (default: "hsa").}

\item{keyType}{The key type for gene identifiers (default: "ENTREZID").}

\item{use_internal_data}{Logical, use internal data for KEGG enrichment (default: FALSE).}

\item{ont}{The ontology to use for GO enrichment (default: "ALL").}

\item{pvalueCutoff}{The p-value cutoff for enrichment (default: 0.05).}

\item{pAdjustMethod}{The method for p-value adjustment (default: "BH").}

\item{universe}{The background gene set.}

\item{qvalueCutoff}{The q-value cutoff for enrichment (default: 0.2).}

\item{minGSSize}{The minimum gene set size for enrichment (default: 10).}

\item{maxGSSize}{The maximum gene set size for enrichment (default: 500).}

\item{readable}{Logical, make the output readable (default: FALSE).}

\item{pool}{Logical, pool genes for GO enrichment (default: FALSE).}

\item{...}{Additional arguments passed to the underlying functions.}
}
\description{
This function is a wrapper for enrich_pathway_internal. It dispatches the call to the
appropriate method based on the class of the object argument. This functions
is based on the clusterProfiler and ReactomePA package.
}
