% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-data_in_package.R
\docType{data}
\name{enriched_functional_module}
\alias{enriched_functional_module}
\title{enriched_functional_module}
\format{
A enriched_functional_module object.
}
\usage{
enriched_functional_module
}
\description{
enriched_functional_module
}
\examples{
data(enriched_functional_module)
}
\keyword{data}
