% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data_in_package.R
\docType{data}
\name{enriched_pathways}
\alias{enriched_pathways}
\title{enriched_pathways}
\format{
An enriched_functional_module object.
}
\usage{
enriched_pathways
}
\description{
enriched_pathways
}
\examples{
data(enriched_pathways)
}
\keyword{data}
