% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\docType{data}
\name{env}
\alias{env}
\title{Compute Semantic Similarity Between GO Terms}
\format{
An object of class \code{environment} of length 0.
}
\source{
Source code download link: https://www.bioconductor.org/packages/release/bioc/src/contrib/simplifyEnrichment_2.0.0.tar.gz
}
\usage{
env
}
\arguments{
\item{go_id}{A character vector of GO term IDs.}

\item{ont}{A character string specifying the ontology.}

\item{db}{A character string specifying the annotation database to use. Default is \code{"org.Hs.eg.db"}.}

\item{measure}{A character string specifying the semantic similarity measure to use. Default is \code{"Sim_XGraSM_2013"}. See \code{simona::all_term_sim_methods()} for available measures.}

\item{control.method}{a list of parameters passing to specified measure method for GO term semantic similarity. For details about how to set this parameter, please go to https://jokergoo.github.io/simona/articles/v05_term_similarity.html.}
}
\value{
A numeric matrix containing the pairwise semantic similarity scores between the provided GO terms.
}
\description{
This function computes the semantic similarity between Gene Ontology (GO) terms using a specified similarity measure. For more information about the methods, please go to https://jokergoo.github.io/simona/articles/v05_term_similarity.html.
}
\note{
This function was adapted from GO_similarity() in the simplifyEnrichment package (version 2.0.0) by Zuguang Gu.
}
\references{
Gu Z, Huebschmann D (2021). “simplifyEnrichment: an R/Bioconductor package for Clustering and Visualizing Functional Enrichment Results.” Genomics, Proteomics & Bioinformatics. doi:10.1016/j.gpb.2022.04.008.
}
\keyword{internal}
