% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/13_data_visualization_plot_module_info_functional_module.R
\name{export_module_info_plot}
\alias{export_module_info_plot}
\title{Export Module Information Plots}
\usage{
export_module_info_plot(
  object,
  path = "result",
  image_type = c("pdf", "png", "jpeg"),
  width = 21,
  height = 7
)
}
\arguments{
\item{object}{An object of class 'functional_module' containing the module information.}

\item{path}{The file path where the plots will be saved.
Default is "result".}

\item{image_type}{The format of the output image, can be 'pdf', 'png', or 'jpeg'.
Default is 'pdf'.}

\item{width}{Width of the output image in inches. Default is 21.}

\item{height}{Height of the output image in inches. Default is 7.}
}
\value{
This function does not return a value but outputs image files to the specified path.
}
\description{
This function exports plots of module information for different databases (GO, KEGG, Reactome, etc.)
from a given 'functional_module' object. It supports multiple image formats.
}
\examples{
\dontrun{
  # Assuming 'my_module' is a valid functional_module object
  export_module_info_plot(object = my_module, path = "my_results", image_type = "png")
}

}
