% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-dplyr_filter.R
\name{filter_functional_module}
\alias{filter_functional_module}
\title{Filter Functional Modules}
\usage{
filter_functional_module(
  object,
  level = c("functional_module", "module", "pathway", "molecule"),
  remain_id
)
}
\arguments{
\item{object}{An object containing data on functional modules, modules, pathways, or molecules.}

\item{level}{A character string specifying the level at which filtering should occur. Valid levels are "functional_module", "module", "pathway", and "molecule".}

\item{remain_id}{A vector of IDs to retain in the filtering process.}
}
\value{
Returns the modified object with filtered data based on the given level and IDs.
}
\description{
This function filters functional modules, modules, pathways, or molecules in the given object based on specified IDs.
}
\note{
This function modifies the input object in-place,
depending on the specified level and IDs to remain.
}
