% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_functional_module_class.R
\docType{class}
\name{functional_module-class}
\alias{functional_module-class}
\title{An S4 class to represent enriched pathways}
\description{
An S4 class to represent enriched pathways
}
\section{Slots}{

\describe{
\item{\code{variable_info}}{variable information contains marker information.}

\item{\code{enrichment_go_result}}{ANY Object storing GO enrichment results.}

\item{\code{enrichment_kegg_result}}{ANY Object storing KEGG enrichment results.}

\item{\code{enrichment_reactome_result}}{ANY Object storing Reactome enrichment results.}

\item{\code{enrichment_hmdb_result}}{ANY Object storing HMDB enrichment results.}

\item{\code{enrichment_metkegg_result}}{ANY Object storing KEGG metabolite enrichment results.}

\item{\code{merged_pathway_go}}{list List containing merged GO pathway information.}

\item{\code{merged_pathway_kegg}}{list List containing merged KEGG pathway information.}

\item{\code{merged_pathway_reactome}}{list List containing merged Reactome pathway information.}

\item{\code{merged_pathway_hmdb}}{list List containing merged HMDB pathway information for metabolite enrichment result.}

\item{\code{merged_pathway_metkegg}}{list List containing merged KEGG pathway information for metabolite enrichment result.}

\item{\code{merged_module}}{list List containing merged modules.}

\item{\code{llm_module_interpretation}}{list List containing LLM interpretation results for functional modules.}

\item{\code{process_info}}{list List containing information about the processes.}
}}

