% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{get_go_result_sim}
\alias{get_go_result_sim}
\title{Pathway Enrichment Analysis and GO Similarity Measurement}
\usage{
get_go_result_sim(
  result,
  sim.cutoff = 0,
  measure.method = "Sim_XGraSM_2013",
  control.method = list()
)
}
\arguments{
\item{result}{A data frame containing GO term IDs and ontologies.}

\item{sim.cutoff}{A numeric value for the similarity cutoff (default: 0).}

\item{measure.method}{A character vector specifying the semantic similarity
measure methods for GO terms. Default is \code{"Sim_XGraSM_2013"}. See \code{simona::term_sim()} for available measures.}

\item{control.method}{a list of parameters passing to specified measure method for GO term semantic similarity. For details about how to set this parameter, please go to https://jokergoo.github.io/simona/articles/v05_term_similarity.html.}
}
\value{
A data frame containing pairs of GO terms and their similarity values.
}
\description{
This function allows for the execution of pathway enrichment analysis and
measurement of semantic similarity among Gene Ontology (GO) terms.
}
\examples{
\dontrun{
# Assuming `result` is your data frame containing GO term IDs and ontologies.
similarity_matrix <- get_go_result_sim(result = result)
}
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
