% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_llm_module_get_pathway_name_desc_pmid.R
\name{get_go_term_info}
\alias{get_go_term_info}
\title{Extract information from GO terms using GO IDs}
\usage{
get_go_term_info(go_ids)
}
\arguments{
\item{go_ids}{A character vector of GO IDs (e.g., "GO:0008150")}
}
\value{
A list of GO term information, where each element contains:
\item{id}{GO ID}
\item{term_name}{Name of the GO term}
\item{term_definition}{Definition of the GO term}
\item{PMID}{Associated PubMed IDs}
}
\description{
This internal function retrieves term name, definition, and PMIDs for given GO IDs
using the QuickGO API.
}
\author{
Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
