% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{get_hmdb_pathways}
\alias{get_hmdb_pathways}
\title{Get HMDB pathway database}
\source{
This function is adapted from the metpath package by Xiaotao Shen
\url{https://github.com/tidymass/metpath/blob/main/R/9_hmdb_database.R}
}
\usage{
get_hmdb_pathways(threads = 3)
}
\arguments{
\item{threads}{Integer. Number of threads to use for parallel processing (default: 3)}
}
\value{
A pathway_database class object containing HMDB pathway information
}
\description{
Retrieve pathway information from the Human Metabolome Database (HMDB/SMPDB).
This is an internal function used by get_met_pathway_db().
}
\references{
Shen, X., Yan, H., Wang, C. et al. TidyMass an object-oriented reproducible analysis framework
for LC–MS data. Nat Commun 13, 4365 (2022).
}
\keyword{internal}
