% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-utils.R
\name{get_jaccard_index_for_three_databases}
\alias{get_jaccard_index_for_three_databases}
\title{get_jaccard_index_for_three_databases Function}
\usage{
get_jaccard_index_for_three_databases(
  object,
  module_result_go,
  module_result_kegg,
  module_result_reactome
)
}
\arguments{
\item{object}{An mass_dataset object containing gene information.}

\item{module_result_go}{A data frame containing module results from GO database.}

\item{module_result_kegg}{A data frame containing module results from KEGG database.}

\item{module_result_reactome}{A data frame containing module results from Reactome database.}
}
\value{
A data frame with the Jaccard index values between all pairs of modules across the three databases.
}
\description{
This function computes the Jaccard index, a measure of similarity for the genes listed in the results of three different databases (GO, KEGG, Reactome).
It first extracts the relevant gene information from the provided object, then computes the Jaccard index between all pairs of modules across the three databases.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
