% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_llm_module_get_pathway_name_desc_pmid.R
\name{get_kegg_info}
\alias{get_kegg_info}
\title{Extract information from KEGG pathways using KEGG IDs}
\usage{
get_kegg_info(kegg_ids)
}
\arguments{
\item{kegg_ids}{A character vector of KEGG pathway IDs (e.g., "hsa04010")}
}
\value{
A list of KEGG pathway information, where each element contains:
\item{id}{KEGG pathway ID}
\item{term_name}{Name of the pathway}
\item{term_definition}{Description of the pathway}
\item{PMID}{Associated PubMed IDs}
}
\description{
This internal function retrieves pathway name, definition, and PMIDs for given KEGG IDs
using the KEGGREST package.
}
\author{
Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
