% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-mapa_logo.R
\name{get_mapa_version}
\alias{get_mapa_version}
\title{Retrieve the Version of mapa Package}
\usage{
get_mapa_version()
}
\value{
A character string indicating the version of the mapa package.
}
\description{
This function retrieves the current version of the mapa package installed in the R environment.
}
\examples{
get_mapa_version()
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
