% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_llm_module_get_pathway_name_desc_pmid.R
\name{get_metkegg_info}
\alias{get_metkegg_info}
\title{Extract metabolic pathway information from KEGG}
\usage{
get_metkegg_info(kegg_ids)
}
\arguments{
\item{kegg_ids}{A character vector of KEGG metabolic pathway IDs}
}
\value{
A list of metabolic pathway information, where each element contains:
\item{id}{KEGG pathway ID}
\item{term_name}{Name of the pathway}
\item{term_definition}{Description of the pathway}
\item{PMID}{Associated PubMed IDs}
}
\description{
This internal function retrieves metabolic pathway names, descriptions, and PMIDs
for given KEGG IDs.
}
\author{
Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
