% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_llm_module_RAG_strategy.R
\name{get_module_embedding}
\alias{get_module_embedding}
\title{Generate Module Embedding}
\usage{
get_module_embedding(
  module_list,
  api_key,
  embedding_model = "text-embedding-3-small",
  api_provider = "openai"
)
}
\arguments{
\item{module_list}{A list or data frame containing module information with at least:
\code{PathwayNames} (character vector) and one of \code{GeneNames_vec} or \code{MetNames_vec}
(character vectors defining molecules of interest).}

\item{api_key}{A character string containing the API key for the embedding generation service.}

\item{embedding_model}{A string specifying the embedding model to use (default is \code{"text-embedding-3-small"}).}

\item{api_provider}{A string indicating the API provider, either \code{"openai"} or \code{"gemini"} (default is \code{"openai"}).}
}
\value{
A numeric vector representing the generated embedding for the module.
}
\description{
Generates an embedding for a given module based on its pathway and gene descriptions.
}
\author{
Feifan Zhang \email{FEIFAN004@e.ntu.edu.sg}
}
\keyword{internal}
