% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_llm_module_get_pathway_name_desc_pmid.R
\name{get_reactome_info}
\alias{get_reactome_info}
\title{Extract information from Reactome pathways using Reactome IDs}
\usage{
get_reactome_info(reactome_ids, orgdb = org.Hs.eg.db)
}
\arguments{
\item{reactome_ids}{A character vector of Reactome pathway IDs}

\item{orgdb}{An OrganismDb object, defaults to org.Hs.eg.db}
}
\value{
A list of Reactome pathway information, where each element contains:
\item{id}{Reactome pathway ID}
\item{term_name}{Name of the pathway}
\item{term_definition}{Description of the pathway}
\item{PMID}{Associated PubMed IDs}
}
\description{
This internal function retrieves pathway name, definition, and PMIDs for given Reactome IDs
using the rbioapi package.
}
\author{
Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
