% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_llm_module_get_pathway_name_desc_pmid.R
\name{get_smpdb_info}
\alias{get_smpdb_info}
\title{Extract pathway information from SMPDB (small molecule pathway database)}
\usage{
get_smpdb_info(smpdb_ids)
}
\arguments{
\item{smpdb_ids}{A character vector of SMPDB pathway IDs}
}
\value{
A list of SMPDB pathway information, where each element contains:
\item{id}{SMPDB pathway ID}
\item{term_name}{Name of the pathway}
\item{term_definition}{Description of the pathway}
\item{PMID}{Associated PubMed IDs (NA for SMPDB)}
}
\description{
This internal function retrieves pathway names, descriptions and related information
for given SMPDB IDs using the metpath package.
}
\author{
Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
