% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_llm_module_get_pathway_name_desc_pmid.R
\name{get_ttl_abstr}
\alias{get_ttl_abstr}
\title{Retrieve titles and abstracts from a list of pathway information}
\usage{
get_ttl_abstr(data)
}
\arguments{
\item{data}{A list of pathway information containing PMIDs}
}
\value{
A list with pathway information and formatted text information including
titles and abstracts from PubMed
}
\description{
This internal function extracts titles and abstracts from PubMed for each PMID
associated with the pathway data.
}
\author{
Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
