% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_merge_modules.R
\name{identify_functional_modules}
\alias{identify_functional_modules}
\title{Identify Functional Modules in a Similarity Matrix}
\usage{
identify_functional_modules(
  query_type = c("gene", "metabolite"),
  variable_info = NULL,
  sim_matrix,
  node_data,
  sim.cutoff = 0.5,
  save_to_local = TRUE,
  path = ".",
  analysis_type = c("enrich_pathway", "do_gsea")
)
}
\arguments{
\item{query_type}{Character, the category of biological entity to query. Must be either "gene" or "metabolite".}

\item{variable_info}{A data frame containing mapping information between different ID types.
For genes, should contain columns: "ensembl", "uniprot", and "entrezid".
For metabolites, should contain columns: "hmdbid" and "keggid".}

\item{sim_matrix}{A data frame containing the similarity matrix with columns \code{name1}, \code{name2}, and \code{value}, representing the similarity between entities.}

\item{node_data}{A data frame containing node information, such as module assignments and other relevant attributes. This parameter cannot be \code{NULL}.}

\item{sim.cutoff}{Numeric. A similarity cutoff value used to filter edges in the similarity matrix. Only edges with similarity values greater than \code{sim.cutoff} will be included in the analysis. Default is \code{0.5}.}

\item{save_to_local}{Logical. Whether to save the resulting data to local files. Default is \code{TRUE}.}

\item{path}{Character. The directory path where the results will be saved, if \code{save_to_local = TRUE}. Default is \code{"."} (the current working directory).}

\item{analysis_type}{Character. Type of analysis to perform: either \code{"enrich_pathway"} or \code{"do_gsea"}. Default is \code{"enrich_pathway"}.}
}
\value{
A list containing:
\item{graph_data}{A \code{tidygraph} object representing the constructed network, including nodes and edges.}
\item{functional_module_result}{A data frame with the identified functional modules and their associated attributes, such as p-values and pathway descriptions.}
\item{result_with_module}{A data frame with the node data enriched with module information.}
}
\description{
This function identifies functional modules from a similarity matrix and node data. It constructs a network graph, applies clustering algorithms, and generates module-level results, which can be saved locally. The function supports both pathway enrichment and gene set enrichment analysis (GSEA).
}
\details{
The function first constructs a graph using the similarity matrix and node data, then applies the edge betweenness clustering algorithm to identify functional modules. The module information is further processed to summarize results such as the number of nodes and enriched pathways for each module. If \code{save_to_local} is \code{TRUE}, the results are saved as files in the specified \code{path}.
}
\examples{
\dontrun{
sim_matrix <-
  data.frame(
    name1 = c("A", "B", "C"),
    name2 = c("B", "C", "A"),
    value = c(0.6, 0.7, 0.8)
  )
node_data <-
  data.frame(module = c("M1", "M2", "M3"),
             p_adjust = c(0.01, 0.05, 0.03))
result <-
  identify_functional_modules(sim_matrix,
                              node_data,
                              sim.cutoff = 0.5,
                              save_to_local = FALSE)
}

}
