% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-llm_intepretation.R
\name{inteprete_pathways}
\alias{inteprete_pathways}
\title{Interpret Enriched Pathways in Relation to a Specific Disease}
\usage{
inteprete_pathways(
  object,
  disease = "pregnancy",
  p.adjust.cutoff = 0.05,
  count.cutoff = 5,
  top_n = 3
)
}
\arguments{
\item{object}{An object of class 'functional_module'. The function will
interpret pathways based on this object.}

\item{disease}{A character string specifying the disease of interest.
Default is "pregnancy".}

\item{p.adjust.cutoff}{A numeric value for the adjusted p-value cutoff
for pathway enrichment. Default is 0.05.}

\item{count.cutoff}{A numeric value specifying the minimum count cutoff
for pathway enrichment. Default is 5.}

\item{top_n}{An integer indicating the number of top pathways to consider
based on the p.adjust.cutoff and count.cutoff. Default is 3.}
}
\value{
A list containing two elements: 'functions' and
'relationship_with_disease'. Each element is a list of character
strings, with each string providing information for a specific
pathway in markdown format.
}
\description{
This function interprets the enriched pathways from an object of class
'functional_module' in relation to a specific disease. It filters pathways
based on p-value and count thresholds and returns their functions and
relationships with a given disease.
}
\details{
The function first checks if the input object is of the correct
class. It then filters the enriched pathways based on the provided
p-value and count thresholds. For each of the top pathways, the
function retrieves information about their functions and their
relationships with the specified disease. This information is
gathered using an external service ('request_chatgpt_response').
}
