% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_functions.R
\name{mapa}
\alias{mapa}
\title{This function computes the one-step estimator from Jones, Gadiyar, and Vandekar (2025) when sample splits are used.
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.}
\usage{
mapa(y, y_hat, folds, y_hat2 = NULL, level = 0.95)
}
\arguments{
\item{y}{A numeric vector of continuous values representing the real outcome values.}

\item{y_hat}{A numeric matrix of continuous values representing the predicted outcome values by subject (row) with each column representing a different sample split.}

\item{folds}{A numeric matrix representing the fold number of the corresponding y_hat and y value (row) with each column representing a different sample split.}

\item{y_hat2}{A numeric matrix of continuous values representing the predicted outcome values by subject (row) from a different model with each column representing a different sample split.}

\item{level}{A number between 0 and 1 representing the confidence level.}
}
\value{
Returns a list with a data frame of the one-step estimator, the logit-scale one-step estimator value, and its lower bound, upper bound, and standard error; a vector of the one-step estimate values by fold for the first model; the sample number; and the confidence level. If y_hat2 is passed as an argument, the one-step estimate for y_hat2 and its difference/ratio to the one-step estimate for y_hat1 are included with confidence intervals.
}
\description{
This function computes the one-step estimator from Jones, Gadiyar, and Vandekar (2025) when sample splits are used.
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.
}
