% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_mapa_logo.R
\name{mapa_logo}
\alias{mapa_logo}
\title{Display the mapa Logo}
\usage{
mapa_logo()
}
\value{
None. This function is called for its side effect of printing the logo and messages.
}
\description{
This function displays the logo and version information for the mapa package.
}
\examples{
mapa_logo()
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
