% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{mapa_packages}
\alias{mapa_packages}
\title{List all packages in the mapa}
\usage{
mapa_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include mapa in the list?}
}
\value{
mapa packages
}
\description{
List all packages in the mapa
}
\examples{
mapa_packages()
}
