% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{mapaFolds_auc}
\alias{mapaFolds_auc}
\title{Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.}
\usage{
mapaFolds_auc(y, y_hat, folds, y_hat2 = NULL, level = 0.95)
}
\arguments{
\item{y}{A numeric vector of binary values representing the real outcome values.}

\item{y_hat}{A numeric vector of binary values representing the predicted outcome values.}

\item{folds}{A numeric vector representing the fold number of the corresponding y_hat and y value.}

\item{y_hat2}{A numeric vector of binary values representing the predicted outcome values from a different model.}

\item{level}{A number between 0 and 1 representing the confidence level.}
}
\value{
Returns a list with a data frame of the one-step estimate for AUC, its lower bound, upper bound, and standard error; a vector of the one-step estimate values by fold for the first model; the sample number; and the confidence level. If y_hat2 is passed as an argument, the one-step estimate for y_hat2 and its difference/ratio to the one-step estimate for y_hat1 are included with confidence intervals.
}
\description{
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.
}
