% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-merge_modules.R
\name{merge_modules}
\alias{merge_modules}
\title{merge_modules Function}
\usage{
merge_modules(
  object,
  module_result_go,
  module_result_kegg,
  module_result_reactome,
  sim.cutoff = 0.5,
  measure_method = c("jaccard"),
  path = "result"
)
}
\arguments{
\item{object}{An mass_dataset object containing gene information.}

\item{module_result_go}{A data frame containing module results from GO database.}

\item{module_result_kegg}{A data frame containing module results from KEGG database.}

\item{module_result_reactome}{A data frame containing module results from Reactome database.}

\item{sim.cutoff}{A numeric value for similarity cutoff. Default is 0.5.}

\item{measure_method}{A character vector specifying the method of similarity measurement. Default is "jaccard".}

\item{path}{A character string specifying the directory to save results. Default is "result".}
}
\value{
Outputs an Excel file and a PDF containing the merged and analyzed data.
}
\description{
This function merges information from three databases (GO, KEGG, Reactome) and performs similarity analysis among pathways/modules across these databases.
It filters, rearranges, and aggregates the data before computing the Jaccard Index for similarity.
The function also creates network visualizations and saves the results to an Excel file.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
