% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-merge_pathways.R
\name{merge_pathways}
\alias{merge_pathways}
\title{Merge Pathways and Analyze Semantic Similarity}
\usage{
merge_pathways(
  pathway_result,
  p.adjust.cutoff = 0.05,
  count.cutoff = 5,
  database = c("go", "kegg", "reactome"),
  sim.cutoff = 0.5,
  measure_method = c("Wang", "Resnik", "Rel", "Jiang", "Lin", "TCSS", "jaccard"),
  path = "result"
)
}
\arguments{
\item{pathway_result}{A result object from pathway enrichment analysis.}

\item{p.adjust.cutoff}{A numeric value for the adjusted p-value cutoff (default: 0.05).}

\item{count.cutoff}{A numeric value for the count cutoff (default: 5).}

\item{database}{A character vector specifying the database (default: c("go", "kegg", "reactome")).}

\item{sim.cutoff}{A numeric value for the similarity cutoff (default: 0.5).}

\item{measure_method}{A character vector specifying the similarity measure method (default: c("Wang", "Resnik", "Rel", "Jiang", "Lin", "TCSS", "jaccard")).}

\item{path}{A character string specifying the path for result output (default: "result").}
}
\value{
NULL. Results are written to files.
}
\description{
This function performs a merging of pathways from enrichment analysis, calculates
semantic similarity among terms, and identifies modules within the pathways.
It also generates various plots and outputs results into organized files.
}
\examples{
\dontrun{
# Assuming `pathway_result` is the result from your pathway enrichment analysis.
merge_pathways(pathway_result = pathway_result)
}
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
