% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_llm_module_output_generation.R
\name{module_name_generation}
\alias{module_name_generation}
\title{Module Name Generation}
\usage{
module_name_generation(
  paper_result,
  phenotype = NULL,
  model = "gpt-4o-mini-2024-07-18",
  api_key,
  output_prompt = TRUE,
  api_provider = "openai",
  thinkingBudget = 0,
  multi_omics = FALSE
)
}
\arguments{
\item{paper_result}{A list containing module information, including related papers and PubMed results.}

\item{phenotype}{Character string. Phenotype or disease to focus on. Default is NULL.}

\item{model}{A string specifying the GPT model to use. Default is \code{"gpt-4o-mini-2024-07-18"}.}

\item{api_key}{A string containing the API key required to access the AI API.}

\item{output_prompt}{Logical. Whether to output prompt in final annotation result. Default is TRUE.}

\item{api_provider}{A string indicating the API provider, either \code{"openai"}, \code{"gemini"}, or \code{"siliconflow"} (default is \code{"openai"}).}

\item{thinkingBudget}{An integer for the "thinking budget" parameter specific to the Gemini API (default is \code{0}).}
}
\value{
A list of results for each module, where each element is a list containing
\code{module_name} and \code{summary}.
}
\description{
This internal function processes a list of modules to generate final biological module names
and summaries for all modules by calling \code{single_module_generation}.
}
\seealso{
\code{\link{single_module_generation}}
}
\author{
Feifan Zhang \email{FEIFAN004@e.ntu.edu.sg}
}
\keyword{internal}
