% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-data_visualization_plot_module_bar.R
\name{plot_module_bar}
\alias{plot_module_bar}
\title{plot_module_bar Function}
\usage{
plot_module_bar(module_result, top_n = 5)
}
\arguments{
\item{module_result}{A data frame containing module results.}

\item{top_n}{An integer specifying the number of top modules to be plotted. Default is 5.}
}
\value{
A ggplot object.
}
\description{
This function generates a bar plot for the top_n modules based on their significance from the module_result data.
It visualizes the -log10 transformed FDR adjusted p-values of the modules alongside their count size.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
