% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-data_visualization_plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot a Network of Modules, Pathways, and Molecules}
\usage{
plot_network(
  module_result = module_result_all,
  result_with_module = result_with_module,
  enrichment_go = enrichment_go,
  enrichment_kegg = enrichment_kegg,
  enrichment_reactome = enrichment_reactome,
  marker_info = marker,
  including_molecule = TRUE,
  circle = FALSE,
  node_color = c(Function_module = "#F05C3BFF", Module = "#46732EFF", Pathway =
    "#197EC0FF", Molecule = "#3B4992FF"),
  text_size = 3,
  arrange_position = TRUE,
  position_ratio = 0.95
)
}
\arguments{
\item{module_result}{A data frame of module results.}

\item{result_with_module}{A data frame of results with module information.}

\item{enrichment_go}{A data frame of GO enrichment results.}

\item{enrichment_kegg}{A data frame of KEGG enrichment results.}

\item{enrichment_reactome}{A data frame of Reactome enrichment results.}

\item{marker_info}{A data frame of marker information.}

\item{including_molecule}{Logical, whether to include molecules in the plot.}

\item{circle}{Logical, whether to use a circular layout for the plot.}

\item{node_color}{A vector of colors for different node types.}

\item{text_size}{Numeric, size of the text in the plot.}

\item{arrange_position}{Logical, whether to arrange position of nodes.}

\item{position_ratio}{Numeric, ratio for arranging position of nodes.}
}
\value{
A \code{ggplot} object representing the network plot.
}
\description{
This function generates a network plot based on the input data which includes module results, pathway enrichments, and marker information.
}
\examples{
\dontrun{
plot_network(module_result,
result_with_module,
enrichment_go,
enrichment_kegg,
enrichment_reactome,
marker_info)
}
}
