% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_data_visualization_plot_pathway_bar.R
\name{plot_pathway_bar}
\alias{plot_pathway_bar}
\title{Plot Enriched Pathway/Module Bar Chart}
\usage{
plot_pathway_bar(
  object,
  top_n = 10,
  x_axis_name = NULL,
  y_label_width = 50,
  translation = FALSE,
  level = c("pathway", "module", "functional_module"),
  line_type = c("straight", "meteor"),
  p.adjust.cutoff = 0.05,
  count.cutoff = 5,
  database_color = c(GO = "#1F77B4FF", KEGG = "#FF7F0EFF", Reactome = "#2CA02CFF", HMDB =
    "#9467BDFF"),
  database = c("go", "kegg", "reactome", "hmdb")
)
}
\arguments{
\item{object}{An object containing the enrichment results and other relevant data.}

\item{top_n}{An integer specifying the top N pathways to display.}

\item{x_axis_name}{A character string indicating the metric to use for the x-axis in ORA results. Options include \code{"qscore"} (i.e. \eqn{-\log10} of the FDR-adjusted p-values), \code{"RichFactor"}, or \code{"FoldEnrichment"}. For metabolite enrichment, the adjusted p-value is used by default, and for GSEA results this parameter is ignored (NES is used instead).}

\item{y_label_width}{An integer specifying the width of the Y-axis labels.}

\item{translation}{translation or not.}

\item{level}{A character string specifying the level of analysis.
One of "pathway", "module", or "functional_module".}

\item{line_type}{A character string specifying the type of line to use for the bar chart.
One of "straight" or "meteor".}

\item{p.adjust.cutoff}{A numeric value for the FDR adjusted P-value cutoff.}

\item{count.cutoff}{A numeric value for the count cutoff.}

\item{database_color}{A named vector containing the colors for different databases.}

\item{database}{A character vector indicating which databases to include in the plot. For gene enrichment analysis valid options are \code{"go"}, \code{"kegg"}, and/or \code{"reactome"}. For metabolite enrichment analysis, valid options are \code{"hmdb"} and/or \code{"kegg"}.}
}
\value{
A ggplot object representing the bar chart.
}
\description{
This function generates a bar chart to visualize enriched pathways, modules, or functional modules.
The bars are colored according to the database of origin.
Depending on the type of enrichment analysis (e.g., ORA or GSEA) and the query type (gene or metabolite), different metrics are displayed on the x-axis.
}
\examples{
\dontrun{
data("enriched_functional_module", package = "mapa")
plot_pathway_bar(
  object = enriched_functional_module,
  top_n = 10,
  y_label_width = 30,
  level = "pathway"
)

plot_pathway_bar(
  object = enriched_functional_module,
  top_n = 10,
  y_label_width = 30,
  level = "pathway",
  line_type = "meteor"
)

plot_pathway_bar(enriched_functional_module,
                 top_n = 10,
                 level = "module")

plot_pathway_bar(
  enriched_functional_module,
  top_n = 10,
  level = "functional_module",
  p.adjust.cutoff = 0.05
)
}
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
