% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-data_visualization_plot_similarity_network.R
\name{plot_similarity_network}
\alias{plot_similarity_network}
\title{Plot Similarity Network}
\usage{
plot_similarity_network(
  object,
  level = c("module", "functional_module"),
  database = c("go", "kegg", "reactome"),
  degree_cutoff = 0,
  module_id,
  text = TRUE,
  text_all = FALSE
)
}
\arguments{
\item{object}{An object of class "functional_module" containing
the necessary data for plotting.}

\item{level}{A character string indicating the level of analysis,
either "module" or "functional_module".}

\item{database}{A character string indicating the database to be used,
either "go", "kegg", or "reactome".}

\item{degree_cutoff}{A numeric value indicating the degree cutoff for
filtering nodes in the network.}

\item{module_id}{Optional, a character or numeric vector of module IDs
to include in the plot.}

\item{text}{Logical, whether to include text annotations on the plot.}

\item{text_all}{Logical, whether to include text annotations for all nodes.}
}
\value{
A ggplot object representing the similarity network.
}
\description{
This function plots a similarity network based on the given parameters.
It can operate on different levels and databases, and provides options for
text annotations.
}
\examples{
\dontrun{
# Assume `obj` is a prepared object containing relevant data
plot_similarity_network(obj, level = "module", database = "go")
plot_similarity_network(obj, level = "functional_module", degree_cutoff = 2)
}

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
