% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_llm_module_online_retrieval.R
\name{pubmed_search}
\alias{pubmed_search}
\title{Search PubMed for All Modules and Add PubMed IDs}
\usage{
pubmed_search(
  processed_data,
  phenotype,
  chunk_size = 5,
  years = 5,
  retmax = 10,
  thread = 10
)
}
\arguments{
\item{processed_data}{A named list where each element corresponds to a module.}

\item{chunk_size}{An integer specifying the size of query chunks (default is 5).}

\item{years}{An integer specifying how many years to look back in the search (default is 5).}

\item{retmax}{An integer specifying the maximum number of results to retrieve (default is 10).}

\item{thread}{An integer specifying the number of parallel threads to use for processing.
Default is \code{10} for sequential processing.}
}
\value{
A named list similar to \code{processed_data}, but with an added \code{PubmedIDs}
field for each module, containing the retrieved PubMed IDs.
}
\description{
This internal function performs a PubMed search for each module in the processed data,
retrieves PubMed IDs based on gene symbols, descriptions, and pathways, and appends
the IDs to the processed data.
}
\author{
Feifan Zhang \email{FEIFAN004@e.ntu.edu.sg}

Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
