% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_llm_module_RAG_strategy.R
\name{read_titles}
\alias{read_titles}
\title{Read Titles}
\usage{
read_titles(save_dir, start_row = 1, n_rows = -1)
}
\arguments{
\item{save_dir}{A character string specifying the output directory where the titles file is stored.}

\item{start_row}{An integer indicating the starting row to read from (default is 1).}

\item{n_rows}{An integer specifying the number of rows to read. Use -1 to read all rows (default is -1).}
}
\value{
A character vector containing the titles.
}
\description{
Reads paper titles from a specified directory. This function supports batch reading of data.
}
\examples{
\dontrun{
# Example: Read the first 50 titles
titles <- read_titles("example_dir", start_row = 1, n_rows = 50)
}

}
\author{
Feifan Zhang \email{FEIFAN004@e.ntu.edu.sg}
}
\keyword{internal}
