% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_report_functional_module.R
\name{report_functional_module}
\alias{report_functional_module}
\alias{report_functional_module.functional_module}
\alias{report_functional_module.list}
\title{Generate a Report for a Functional Module}
\usage{
report_functional_module(object, path = ".", ...)

\method{report_functional_module}{functional_module}(
  object,
  path = ".",
  degree_cutoff = 1,
  type = c("html", "pdf", "word", "md", "all"),
  ...
)

\method{report_functional_module}{list}(object, path = ".", ...)
}
\arguments{
\item{object}{An object of class 'functional_module'. The function will
generate a report based on this object.}

\item{path}{A character string specifying the directory path where the report
and associated files will be saved. Defaults to the current directory.}

\item{degree_cutoff}{Minimum node degree for the similarity network plot
(default 1).}

\item{type}{A character vector specifying the format of the report.
Can be 'html', 'pdf', 'md', 'word', or 'all' to generate the report
in all three formats. Default is 'html'.}
}
\value{
This function does not return a value but generates a report in the
specified format(s) and saves it to the given path.
}
\description{
This function generates a report for a given object of class 'functional_module'.
It creates the report in different formats (HTML, PDF, Word) based on user choice.
The function also performs various tasks such as creating directories,
fetching report templates, saving parameter and object data,
and rendering bar plots for functional modules, modules, and pathways, similarity networks,
and interpretation of functional modules.
}
\details{
The function first checks for the presence and class of the 'object'
parameter. It then creates the necessary directory structure, fetches
the report template using the 'draft' function from the 'rmarkdown' package,
and saves parameters and object data. It also creates and saves bar plots
for different levels (functional_module, module, pathway) using 'ggplot2'.
Finally, it renders the report in the specified format(s) and cleans up
intermediate files.
}
