% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/98_openai_chatgpt.R
\name{request_chatgpt_response}
\alias{request_chatgpt_response}
\title{Request Response from ChatGPT}
\usage{
request_chatgpt_response(prompt = "what is metabolomics?")
}
\arguments{
\item{prompt}{A string representing the prompt to be sent to the ChatGPT service.
Default prompt is "what is metabolomics?".}
}
\value{
A string containing the response from ChatGPT.
If the API key is not set, the function stops with an error message.
}
\description{
This function sends a request to OpenAI's ChatGPT service using the provided prompt.
It requires that a ChatGPT API key is already set using \code{set_chatgpt_api_key()}.
The function makes a POST request to the ChatGPT API and returns the generated response.
}
\examples{
\dontrun{
  # Set the ChatGPT API key first
  set_chatgpt_api_key("your_api_key_here")

  # Request a response from ChatGPT service
  chatgpt_response <- request_chatgpt_response(prompt = "Tell me about R programming")
}

}
