% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_google_gemini.R
\name{request_gemini_response}
\alias{request_gemini_response}
\title{Request Response from Gemini Service}
\usage{
request_gemini_response(prompt = "what is metabolomics?")
}
\arguments{
\item{prompt}{A string representing the prompt to be sent to the Gemini service.}
}
\value{
A character vector containing the response outputs from the Gemini service.
If the API key is not set, the function stops with an error message.
}
\description{
This function sends a request to the Gemini service using the provided prompt.
It requires that a Gemini API key is already set using \code{set_gemini_api_key()}.
The function makes a POST request and returns the response from Gemini.
}
\examples{
\dontrun{
  # Set the API key first
  set_gemini_api_key("your_api_key_here")

  # Request a response from Gemini service
  response <- request_gemini_response("Your prompt here")
}

}
