% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/98-openai_chatgpt.R
\name{set_chatgpt_api_key}
\alias{set_chatgpt_api_key}
\title{Set ChatGPT API Key}
\usage{
set_chatgpt_api_key(api_key)
}
\arguments{
\item{api_key}{A string representing the ChatGPT API key to be set.}
}
\value{
Invisible NULL. The function is used for its side effect
of setting an environment variable and does not return a value.
}
\description{
This function sets the ChatGPT API key as an environment variable.
It's used to authenticate and interact with ChatGPT API services.
}
\examples{
\dontrun{
# Set the ChatGPT API key (example key used here)
set_chatgpt_api_key("your_api_key_here")
}
}
