% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-google_gemini.R
\name{set_gemini_api_key}
\alias{set_gemini_api_key}
\title{Set Gemini API Key}
\usage{
set_gemini_api_key(api_key)
}
\arguments{
\item{api_key}{A string representing the Gemini API key to be set.}
}
\value{
Invisible NULL. The function is used for its side effect
of setting an environment variable and does not return a value.
}
\description{
This function sets the Gemini API key as an environment variable.
It's used to authenticate and interact with Gemini API services.
}
\examples{
\dontrun{
# Set the Gemini API key (example key used here)
set_gemini_api_key("your_api_key_here")
}
}
