% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{term_similarity_internal}
\alias{term_similarity_internal}
\title{Similarity calculation between pathways.}
\source{
Source code download link: https://bioconductor.org/packages/3.19/bioc/src/contrib/Archive/simplifyEnrichment/simplifyEnrichment_1.14.0.tar.gz
}
\usage{
term_similarity_internal(
  gl,
  measure.method = c("jaccard", "dice", "overlap", "kappa"),
  remove_negative = TRUE
)
}
\arguments{
\item{gl}{Named list, genes that are in the enriched pathways.}

\item{measure.method}{Character, method for calculating the semantic similarity
for KEGG and Reactome terms}

\item{remove_negative}{Logical, if TRUE reset the negative similarity values to zero}
}
\value{
A symmetric matrix.
}
\description{
This function allows for the execution of measurement of semantic similarity
among KEGG and Reactome terms based on the overlap of genes.
}
\note{
This function was adapted from term_similarity() in the simplifyEnrichment package (version 1.14.0) by Zuguang Gu.
}
\references{
Gu Z, Huebschmann D (2021). “simplifyEnrichment: an R/Bioconductor package for Clustering and Visualizing Functional Enrichment Results.” Genomics, Proteomics & Bioinformatics. doi:10.1016/j.gpb.2022.04.008.
}
\keyword{internal}
