% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{term_similarity_KEGG}
\alias{term_similarity_KEGG}
\title{Similarity calculation between KEGG pathways}
\source{
Source code download link: https://bioconductor.org/packages/3.19/bioc/src/contrib/Archive/simplifyEnrichment/simplifyEnrichment_1.14.0.tar.gz
}
\usage{
term_similarity_KEGG(
  term_id,
  measure.method = c("jaccard", "dice", "overlap", "kappa")
)
}
\arguments{
\item{term_id}{Character, KEGG pathway IDs}

\item{measure.method}{Character, method for calculating the similarity
for KEGG terms, Choices are "jaccard", "dice", "overlap" and "kappa". Default is "jaccard".}
}
\value{
A symmetric matrix
}
\description{
Similarity calculation between KEGG pathways
}
\note{
This function was adapted from term_similarity__from_KEGG() in the simplifyEnrichment package (version 1.14.0) by Zuguang Gu.
}
\references{
Gu Z, Huebschmann D (2021). “simplifyEnrichment: an R/Bioconductor package for Clustering and Visualizing Functional Enrichment Results.” Genomics, Proteomics & Bioinformatics. doi:10.1016/j.gpb.2022.04.008.
}
