% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{term_similarity_metabolite}
\alias{term_similarity_metabolite}
\title{Similarity calculation between enriched HMDB or KEGG pathways from metabolite enrichment analysis}
\source{
Source code download link: https://bioconductor.org/packages/3.19/bioc/src/contrib/Archive/simplifyEnrichment/simplifyEnrichment_1.14.0.tar.gz
}
\usage{
term_similarity_metabolite(
  enrichment_result,
  measure.method = c("jaccard", "dice", "overlap", "kappa")
)
}
\arguments{
\item{enrichment_result}{Dataframe, result from metabolite enrichment analysis}

\item{measure.method}{Character, method for calculating the similarity
between enriched pathways, Choices are "jaccard", "dice", "overlap", "kappa". Default is "jaccard".}
}
\value{
A symmetric matrix
}
\description{
Similarity calculation between enriched HMDB or KEGG pathways from metabolite enrichment analysis
}
\note{
This function was adapted from the simplifyEnrichment package (version 1.14.0) by Zuguang Gu.
}
\references{
Gu Z, Huebschmann D (2021). “simplifyEnrichment: an R/Bioconductor package for Clustering and Visualizing Functional Enrichment Results.” Genomics, Proteomics & Bioinformatics. doi:10.1016/j.gpb.2022.04.008.
}
