% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100-translation.R
\name{translate_language}
\alias{translate_language}
\title{Translate Text into Different Languages}
\usage{
translate_language(
  text =
    "As of my last update in early 2023, there's no widely\\n           recognized individual named Xiaotao Shen affiliated with Stanford\\n           University that I can refer to.\\n           It is important to note that personal details about\\n           researchers or students, unless they are part of a widely\\n           recognized work or public academic profile,\\n           may not be available in the public domain.\\n           If Xiaotao Shen is a researcher, student,\\n           or faculty member at Stanford, you might find\\n           information on university directories, academic publications,\\n           or professional networking sites such as LinkedIn.\\n           However, for the most updated and accurate information,\\n           it is best to directly consult resources at Stanford University\\n           or other official channels.",
  engine = c("chatgpt", "gemini"),
  to = c("chinese", "spanish", "english", "french", "german", "italian", "japanese",
    "korean", "portuguese", "russian", "spanish")
)
}
\arguments{
\item{text}{A string representing the text to be translated.
Default is a description of the Gemini model.}

\item{engine}{A string specifying the translation engine to be used, "chatgpt" or "gemini".}

\item{to}{A string indicating the target language for translation.
Supported languages include 'chinese', 'spanish', 'english', 'french', 'german',
'italian', 'japanese', 'korean', 'portuguese', 'russian', and 'spanish'.}
}
\value{
A character vector containing the translated text.
}
\description{
This function translates a given text into a specified language using the Gemini translation engine.
It utilizes the Gemini large language models for translation.
}
\examples{
\dontrun{
  # Translate text to French using Gemini
  translated_text <- translate_language(text = "Hello World", engine = "gemini", to = "french")
}

}
