% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mnetwork_class.R
\name{build_MNetwork}
\alias{build_MNetwork}
\title{Build a multi_omics_functional_module object}
\usage{
build_MNetwork(
  network_tables,
  mol_layers_m = list(),
  path_adj = NULL,
  layer_weights = NULL,
  params = list()
)
}
\arguments{
\item{network_tables}{\code{list}. Standardised network tables.}

\item{mol_layers_m}{\code{list}. Edge weight matrix for each molecule layers.}

\item{path_adj}{\code{dMatrix} or \code{NULL}. m × m pathway adjacency matrix.
A zero matrix is used when \code{NULL}.}

\item{layer_weights}{\code{numeric} or \code{NULL}. Weight vector of length L.
Equal weights (1/L) are assigned when \code{NULL}.}

\item{params}{\code{list}. Named list overriding default diffusion parameters:
\code{restart}, \code{interlayer}, \code{gamma}, \code{tol}, \code{max_iter}.}
}
\value{
A validated \link{multi_omics_functional_module} object.
}
\description{
Constructs a \link{multi_omics_functional_module} object.
}
