% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/15_data_visualization_plot_relationship_network.R
\name{create_relation_network}
\alias{create_relation_network}
\title{Create a Relationship Network}
\usage{
create_relation_network(
  object,
  include_functional_modules = TRUE,
  include_modules = TRUE,
  include_pathways = TRUE,
  include_molecules = TRUE,
  translation = FALSE
)
}
\arguments{
\item{object}{An object of class \code{functional_module} containing the functional module data.}

\item{include_functional_modules}{Logical; whether to include functional modules in the network. Defaults to TRUE.}

\item{include_modules}{Logical; whether to include modules in the network. Defaults to TRUE.}

\item{include_pathways}{Logical; whether to include pathways in the network. Defaults to TRUE.}

\item{include_molecules}{Logical; whether to include molecules in the network. Defaults to TRUE.}

\item{translation}{translation or not.}
}
\value{
A list containing \code{edge_data} and \code{node_data} data frames.
\itemize{
\item \code{edge_data} is a data frame containing the relationships between nodes, including their classes.
\item \code{node_data} is a data frame containing information about each node, such as annotation and class.
}
}
\description{
This function constructs a network of relationships between different biological entities.
The network can include functional modules, modules, pathways, and molecules.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
