% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_do_gesa.R
\name{do_gsea}
\alias{do_gsea}
\title{Perform Gene Set Enrichment Analysis (GSEA)}
\usage{
do_gsea(
  variable_info,
  query_type = "gene",
  order_by = "fc",
  database = c("go", "kegg", "reactome"),
  save_to_local = FALSE,
  path = "result",
  go.orgdb = NULL,
  go.keytype = "ENTREZID",
  go.ont = "ALL",
  kegg.organism = "hsa",
  kegg.keytype = "kegg",
  reactome.organism = "human",
  exponent = 1,
  eps = 1e-10,
  verbose = TRUE,
  seed = FALSE,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  ...
)
}
\arguments{
\item{variable_info}{A data frame containing variable information.}

\item{query_type}{Character, the category of biological entity to query. Must be one of:
\itemize{
\item "gene": for gene/protein-based enrichment
\item "metabolite": for metabolite-based enrichment (currently unusable)
}}

\item{order_by}{Column name to order genes by. Default is "fc".}

\item{database}{Databases to use: "go", "kegg", "reactome".}

\item{save_to_local}{Save results locally. Default is FALSE.}

\item{path}{Directory to save results. Default is "result".}

\item{go.orgdb}{OrgDb object for GO analysis.}

\item{go.keytype}{Key type for GO. Default is "ENTREZID".}

\item{go.ont}{GO ontology to use. Default is "ALL".}

\item{kegg.organism}{Organism code for KEGG. Default is "hsa".}

\item{kegg.keytype}{Key type for KEGG. Default is "kegg".}

\item{reactome.organism}{Organism for Reactome. Default is "human".
Supported values: "human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly".}

\item{exponent}{GSEA ranking exponent. Default is 1.}

\item{eps}{Tolerance value. Default is 1e-10.}

\item{verbose}{Print messages. Default is TRUE.}

\item{seed}{Set random seed. Default is FALSE.}

\item{pvalueCutoff}{P-value cutoff. Default is 0.05.}

\item{pAdjustMethod}{P-value adjustment method. Default is "BH".}

\item{qvalueCutoff}{Q-value cutoff. Default is 0.2.}

\item{minGSSize}{Minimum gene set size. Default is 10.}

\item{maxGSSize}{Maximum gene set size. Default is 500.}

\item{readable}{Convert IDs to symbols. Default is FALSE.}

\item{...}{Additional arguments}
}
\value{
A functional_module object containing enrichment results
}
\description{
Perform Gene Set Enrichment Analysis (GSEA)
}
