% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_enrich_pathway.R
\name{enrich_pathway}
\alias{enrich_pathway}
\title{Enrich Pathways}
\usage{
enrich_pathway(
  variable_info,
  query_type = c("gene", "metabolite"),
  database = c("go", "kegg", "reactome", "hmdb"),
  save_to_local = FALSE,
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe_go = NULL,
  universe_kegg = NULL,
  universe_reactome = NULL,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)
}
\arguments{
\item{variable_info}{A data.frame containing relevant gene information or
an object that has a specific method for \code{enrich_pathway}.}

\item{query_type}{Character, the category of biological entity to query ("gene", "metabolite").}

\item{database}{Character vector, specify which database(s) to use for enrichment ('go', 'kegg', 'reactome', 'hmdb').}

\item{save_to_local}{Logical, if TRUE the results will be saved to local disk.}

\item{path}{Character, the directory where to save the results if save_to_local is TRUE.}

\item{OrgDb}{Object, the OrgDb object required for GO enrichment.}

\item{organism}{Character, the organism code.}

\item{keyType}{Character, the type of key to be used.}

\item{use_internal_data}{Logical, whether to use internal database for KEGG and HMDB enrichment.}

\item{ont}{Character, the ontology to be used for GO enrichment ('ALL', 'BP', 'CC', 'MF').}

\item{pvalueCutoff}{Numeric, the p-value cutoff for enrichment.}

\item{pAdjustMethod}{Character, the method for adjusting p-values.}

\item{universe_go}{Numeric vector, the universe for GO enrichment.}

\item{universe_kegg}{Numeric vector, the universe for KEGG enrichment.}

\item{universe_reactome}{Numeric vector, the universe for Reactome enrichment.}

\item{qvalueCutoff}{Numeric, the q-value cutoff for enrichment.}

\item{minGSSize}{Numeric, the minimum gene set size for enrichment.}

\item{maxGSSize}{Numeric, the maximum gene set size for enrichment.}

\item{readable}{Logical, whether to make the results more human-readable.}

\item{pool}{Logical, whether to use clusterProfiler's pooling feature.}

\item{...}{Additional arguments passed to the underlying methods.}
}
\value{
An object containing the enrichment results and parameters.
}
\description{
Performs pathway enrichment analysis using multiple databases (GO, KEGG, Reactome, HMDB). This
function acts as a generic interface that dispatches to specific methods based on the type of query: gene-based or metabolite-based.
For gene queries, GO, KEGG, and Reactome enrichment are performed;
for metabolite queries, HMDB and KEGG enrichment are performed.
}
\examples{
\dontrun{
data(demo_data)

variable_info <-
  demo_data \%>\%
  massdataset::activate_mass_dataset(what = "variable_info") \%>\%
  dplyr::filter(fdr < 0.05 & score > 0) \%>\%
  massdataset::extract_variable_info()
require(org.Hs.eg.db)
enriched_pathways <-
  enrich_pathway(
    variable_info = variable_info,
    save_to_local = FALSE,
    path = "result",
    OrgDb = org.Hs.eg.db,
    organism = "hsa",
    database = c("kegg", "reactome"),
    ont = "ALL",
    pvalueCutoff = 0.05,
    pAdjustMethod = "BH",
    qvalueCutoff = 0.2,
    minGSSize = 10,
    maxGSSize = 500,
    readable = FALSE,
    pool = FALSE
  )
  }
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
