% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-enrich_pathway.R
\name{enrich_pathway}
\alias{enrich_pathway}
\alias{enrich_pathway.mass_dataset}
\alias{enrich_pathway.data.frame}
\title{Enrich Pathways}
\usage{
enrich_pathway(
  object,
  database = c("go", "kegg", "reactome"),
  save_to_local = FALSE,
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe_go = NULL,
  universe_kegg = NULL,
  universe_reactome = NULL,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)

\method{enrich_pathway}{mass_dataset}(
  object,
  database = c("go", "kegg", "reactome"),
  save_to_local = FALSE,
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe_go = NULL,
  universe_kegg = NULL,
  universe_reactome = NULL,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)

\method{enrich_pathway}{data.frame}(
  object,
  database = c("go", "kegg", "reactome"),
  save_to_local = FALSE,
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe_go = NULL,
  universe_kegg = NULL,
  universe_reactome = NULL,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)
}
\arguments{
\item{object}{A list containing relevant gene information or an object that has a specific method for \code{enrich_pathway}.}

\item{database}{Character vector, specify which database(s) to use for enrichment ('go', 'kegg', 'reactome').}

\item{save_to_local}{Logical, if TRUE the results will be saved to local disk.}

\item{path}{Character, the directory where to save the results if save_to_local is TRUE.}

\item{OrgDb}{Object, the OrgDb object required for GO enrichment.}

\item{organism}{Character, the organism code.}

\item{keyType}{Character, the type of key to be used.}

\item{use_internal_data}{Logical, whether to use internal data for KEGG enrichment.}

\item{ont}{Character, the ontology to be used for GO enrichment ('ALL', 'BP', 'CC', 'MF').}

\item{pvalueCutoff}{Numeric, the p-value cutoff for enrichment.}

\item{pAdjustMethod}{Character, the method for adjusting p-values.}

\item{universe_go}{Numeric vector, the universe for GO enrichment.}

\item{universe_kegg}{Numeric vector, the universe for KEGG enrichment.}

\item{universe_reactome}{Numeric vector, the universe for Reactome enrichment.}

\item{qvalueCutoff}{Numeric, the q-value cutoff for enrichment.}

\item{minGSSize}{Numeric, the minimum gene set size for enrichment.}

\item{maxGSSize}{Numeric, the maximum gene set size for enrichment.}

\item{readable}{Logical, whether to make the results more human-readable.}

\item{pool}{Logical, whether to use clusterProfiler's pooling feature.}

\item{...}{Additional arguments passed to the underlying methods.}
}
\value{
An object containing the enrichment results and parameters.
}
\description{
This function performs enrichment analysis on various databases such as GO, KEGG, and Reactome.
It's a generic function that dispatches to specific methods based on the class of the 'object' parameter.
}
\examples{
\dontrun{
# Create a sample object
sample_object <- list(ensembl = c("ENSG000001", "ENSG000002"), uniprot = c("P12345", "P67890"))

# Perform enrichment
result <- enrich_pathway(object = sample_object, database = c("go", "kegg"), OrgDb = org.Hs.eg.db)
}
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
