% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-enrich_pathway.R
\name{enrich_pathway_internal}
\alias{enrich_pathway_internal}
\title{Internal Function for Enriching Pathways}
\usage{
enrich_pathway_internal(
  object,
  database = c("go", "kegg", "reactome"),
  save_to_local = FALSE,
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe_go = NULL,
  universe_kegg = NULL,
  universe_reactome = NULL,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)
}
\arguments{
\item{object}{A list containing relevant gene information.}

\item{database}{Character vector, specify which database(s) to use for enrichment ('go', 'kegg', 'reactome').}

\item{save_to_local}{Logical, if TRUE the results will be saved to local disk.}

\item{path}{Character, the directory where to save the results if save_to_local is TRUE.}

\item{OrgDb}{Object, the OrgDb object required for GO enrichment.}

\item{organism}{Character, the organism code.}

\item{keyType}{Character, the type of key to be used.}

\item{use_internal_data}{Logical, whether to use internal data for KEGG enrichment.}

\item{ont}{Character, the ontology to be used for GO enrichment ('ALL', 'BP', 'CC', 'MF').}

\item{pvalueCutoff}{Numeric, the p-value cutoff for enrichment.}

\item{pAdjustMethod}{Character, the method for adjusting p-values.}

\item{universe_go}{Numeric vector, the universe for GO enrichment.}

\item{universe_kegg}{Numeric vector, the universe for KEGG enrichment.}

\item{universe_reactome}{Numeric vector, the universe for Reactome enrichment.}

\item{qvalueCutoff}{Numeric, the q-value cutoff for enrichment.}

\item{minGSSize}{Numeric, the minimum gene set size for enrichment.}

\item{maxGSSize}{Numeric, the maximum gene set size for enrichment.}

\item{readable}{Logical, whether to make the results more human-readable.}

\item{pool}{Logical, whether to use clusterProfiler's pooling feature.}

\item{...}{Additional arguments passed to the underlying functions.}
}
\value{
An object of class 'functional_module' containing the enrichment results and parameters.
}
\description{
This is an internal function that performs enrichment analysis on various databases such as GO, KEGG, and Reactome.
The function takes in a list of genes and additional parameters to perform the enrichment.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
