% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-export_functional_module.R
\name{export_functional_module}
\alias{export_functional_module}
\title{Export Functional Module Analysis Results to Excel}
\usage{
export_functional_module(object, path = "result")
}
\arguments{
\item{object}{An object of class \code{functional_module}. This object should contain
enrichment results and other relevant data.}

\item{path}{The directory path where the resulting Excel file will be saved.
Defaults to "result".}
}
\value{
Saves an Excel file to the specified path.
The Excel file contains sheets for GO, KEGG, and Reactome enrichment results
for pathways and modules, as well as enriched functional modules.
}
\description{
This function exports the results of functional module analysis to an Excel file.
The function takes an object of class \code{functional_module} and writes multiple
sheets corresponding to various types of enrichment results.
}
\author{
Xiaotao Shen, \email{shenxt1990@outloo.com}
}
