% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-data_visualization_plot_module_info.R
\name{export_module_info_plot}
\alias{export_module_info_plot}
\title{Export Module Information Plot}
\usage{
export_module_info_plot(object, path = "result")
}
\arguments{
\item{object}{A functional_module object.}

\item{path}{A character string specifying the directory path to save the plots.}
}
\value{
Prints messages indicating the progress and saves plots in the specified directory.
}
\description{
This function exports plots for functional modules, saving them as PDF files in the specified directory.
Plots for modules from the GO, KEGG, Reactome, and custom functional modules are supported.
}
\examples{
\dontrun{
obj <- functional_module$new(...)
export_module_info_plot(obj, path = "result")
}

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
