% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-utils.R
\name{get_go_result_sim}
\alias{get_go_result_sim}
\title{Pathway Enrichment Analysis and GO Similarity Measurement}
\usage{
get_go_result_sim(
  result,
  sim.cutoff = 0,
  measure.method = c("Wang", "Resnik", "Rel", "Jiang", "Lin", "TCSS")
)
}
\arguments{
\item{result}{A data frame containing GO term IDs and ontologies.}

\item{sim.cutoff}{A numeric value for the similarity cutoff (default: 0).}

\item{measure.method}{A character vector specifying the semantic similarity measure method
(default: c("Wang", "Resnik", "Rel", "Jiang", "Lin", "TCSS")).}
}
\value{
A data frame containing pairs of GO terms and their similarity values.
}
\description{
This function allows for the execution of pathway enrichment analysis and
measurement of semantic similarity among Gene Ontology (GO) terms.
}
\examples{
\dontrun{
# Assuming `result` is your data frame containing GO term IDs and ontologies.
similarity_matrix <- get_go_result_sim(result = result)
}
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
