% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{get_go_result_sim}
\alias{get_go_result_sim}
\title{Calculate similarity among Gene Ontology (GO) terms}
\usage{
get_go_result_sim(
  result,
  go.orgdb = NULL,
  measure.method = "Sim_XGraSM_2013",
  control.method = list()
)
}
\arguments{
\item{result}{A data frame containing GO term IDs in column 'ID' and ontologies in column 'ONTOLOGY'.}

\item{go.orgdb}{An organism-specific database for GO annotations (default: NULL).}

\item{measure.method}{A character vector specifying the semantic similarity
measure method for GO terms. Default is \code{"Sim_XGraSM_2013"}.
See \code{simona::all_term_sim_methods()} for available measures.}

\item{control.method}{A list of parameters passed to the specified measure method for GO term semantic similarity.
For details, see https://jokergoo.github.io/simona/articles/v05_term_similarity.html.}
}
\value{
A data frame containing pairs of GO terms (name1, name2) and their similarity values (sim).
Also includes an attribute "obsolete_terms" listing any obsolete GO terms encountered.
}
\description{
Internal function that computes similarity between GO terms from enrichment results.
This function processes GO terms by ontology category (BP, MF, CC) and calculates pairwise similarities.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}

Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
