% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_utils.R
\name{get_met_pathway_db}
\alias{get_met_pathway_db}
\title{Retrieve Metabolic Pathway Databases}
\usage{
get_met_pathway_db(use_internal_data = TRUE)
}
\arguments{
\item{use_internal_data}{Logical. Indicates whether to use internal (local) data for the KEGG pathway retrieval.
Note that the HMDB pathway retrieval currently does not support online access.}
}
\value{
A named list with two elements:
\describe{
\item{hmdb_pathway}{A list of HMDB pathways.}
\item{kegg_pathway}{A list of KEGG pathways.}
}
}
\description{
This function retrieves metabolic pathway data from the HMDB and KEGG databases.
}
\examples{
\dontrun{
  # Retrieve pathway databases using internal data for KEGG
  pathways <- get_met_pathway_db(use_internal_data = TRUE)

  # Access HMDB pathways
  hmdb <- pathways$hmdb_pathway

  # Access KEGG pathways
  kegg <- pathways$kegg_pathway
}

}
\keyword{internal}
