% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_llm_interpret_module.R
\name{llm_interpret_module}
\alias{llm_interpret_module}
\title{Interpret Functional Module using LLM Integrated with RAG Strategy}
\usage{
llm_interpret_module(
  object,
  api_key,
  embedding_output_dir,
  local_corpus = FALSE,
  local_corpus_dir = NULL,
  save_dir_local_corpus_embed = NULL,
  phenotype = NULL,
  chunk_size = 5,
  years = 5,
  retmax = 10,
  similarity_filter_num = 20,
  GPT_filter_num = 5,
  orgdb = org.Hs.eg.db
)
}
\arguments{
\item{object}{A functional_module class object.}

\item{api_key}{Character string. API key for OpenAI or other embedding service. (Currently, only API key for OpenAI can be used)}

\item{embedding_output_dir}{Character string. Directory where embedding results will be saved.}

\item{local_corpus}{Logical. Whether to use local files. Default is FALSE.}

\item{local_corpus_dir}{Character string. Directory containing local files provided by users.
Required if local_corpus is TRUE.}

\item{save_dir_local_corpus_embed}{Character string. Directory to save embedded local corpus.
Required if local_corpus is TRUE.}

\item{phenotype}{Character string. Phenotype or disease to focus on. Default is NULL.}

\item{chunk_size}{Integer. Chunk size for processing data. Default is 5.}

\item{years}{Integer. Number of recent years to search in PubMed. Default is 5.}

\item{retmax}{Integer. Maximum number of records to return from PubMed. Default is 10.}

\item{similarity_filter_num}{Integer. Number of papers to filter based on similarity. Default is 20.}

\item{GPT_filter_num}{Integer. Number of papers to filter using GPT. Default is 5.}

\item{orgdb}{Object. Organism database for gene annotation, default is org.Hs.eg.db. Only used for gene enrichment results.}
}
\value{
A list containing the final results with module names and study summaries.
}
\description{
This function processes functional module results by retrieving
relevant papers using a Retrieval-Augmented Generation (RAG) strategy. It includes pathway
description extraction, PubMed searching, embedding of search results, and retrieving
related papers with ranking. Finally, input the functional module information and
abstract and title of retrieved and filtered papers into LLM to generate a name and a summary
for each functional module.
}
\author{
Feifan Zhang \email{FEIFAN004@e.ntu.edu.sg}

Yifei Ge \email{yifeii.ge@outlook.com}
}
