% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_merge_modules.R
\name{merge_modules}
\alias{merge_modules}
\title{Identify Functional Modules Across Different Databases}
\usage{
merge_modules(
  object,
  sim.cutoff = 0.5,
  measure_method = c("jaccard"),
  path = "result",
  save_to_local = TRUE
)
}
\arguments{
\item{object}{An S4 object, expected to be processed by \code{merge_pathways()}.}

\item{sim.cutoff}{A numerical value for similarity cutoff, default is 0.5.}

\item{measure_method}{A character vector indicating the method for measuring similarity, default is "jaccard".}

\item{path}{A character string for the directory where to save results, default is "result".}

\item{save_to_local}{Logical, if TRUE the results will be saved to local disk.}
}
\description{
This function identify functional modules from various databases, such as GO, KEGG, and Reactome.
It calculates the similarity matrix between all the pathways and clusters them into functional modules.
}
\note{
Please ensure that \code{object} has been processed by \code{merge_pathways()}
before using this function.
}
\examples{
\dontrun{
merge_modules(variable_info = my_var_info, object = my_object)
}

}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
