% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-merge_pathways.R
\name{merge_pathways}
\alias{merge_pathways}
\title{Merge Pathways from Multiple Databases}
\usage{
merge_pathways(
  object,
  p.adjust.cutoff.go = 0.05,
  p.adjust.cutoff.kegg = 0.05,
  p.adjust.cutoff.reactome = 0.05,
  count.cutoff.go = 5,
  count.cutoff.kegg = 5,
  count.cutoff.reactome = 5,
  sim.cutoff.go = 0.5,
  sim.cutoff.kegg = 0.5,
  sim.cutoff.reactome = 0.5,
  measure.method.go = c("Wang", "Resnik", "Rel", "Jiang", "Lin", "TCSS", "jaccard"),
  measure.method.kegg = c("jaccard"),
  measure.method.reactome = c("jaccard"),
  path = "result",
  save_to_local = FALSE
)
}
\arguments{
\item{object}{An object of class "functional_module", typically a result from enrich_pathway function.}

\item{p.adjust.cutoff.go}{Adjusted p-value cutoff for GO database. Default is 0.05.}

\item{p.adjust.cutoff.kegg}{Adjusted p-value cutoff for KEGG database. Default is 0.05.}

\item{p.adjust.cutoff.reactome}{Adjusted p-value cutoff for Reactome database. Default is 0.05.}

\item{count.cutoff.go}{Count cutoff for GO database. Default is 5.}

\item{count.cutoff.kegg}{Count cutoff for KEGG database. Default is 5.}

\item{count.cutoff.reactome}{Count cutoff for Reactome database. Default is 5.}

\item{sim.cutoff.go}{Similarity cutoff for GO database. Default is 0.5.}

\item{sim.cutoff.kegg}{Similarity cutoff for KEGG database. Default is 0.5.}

\item{sim.cutoff.reactome}{Similarity cutoff for Reactome database. Default is 0.5.}

\item{measure.method.go}{A character vector specifying the similarity measure method for GO. Choices are "Wang", "Resnik", "Rel", "Jiang", "Lin", "TCSS", "jaccard". Default is "Wang".}

\item{measure.method.kegg}{A character vector specifying the similarity measure method for KEGG. Default is "jaccard".}

\item{measure.method.reactome}{A character vector specifying the similarity measure method for Reactome. Default is "jaccard".}

\item{path}{Directory path to save the results. Default is "result".}

\item{save_to_local}{Logical, if TRUE the results will be saved to local disk.}
}
\value{
An object of class "functional_module" with slots for merged pathways from each database.
}
\description{
This function merges pathway enrichment results from multiple databases (GO, KEGG, and Reactome)
into a single object. The function takes an object of class "functional_module" and various
parameters to filter and compute similarity measures.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
