% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_merge_pathways.R
\name{merge_pathways}
\alias{merge_pathways}
\title{This function merges enrichment analysis results from different databases (GO, KEGG, and Reactome) into a single object respectively. The function takes an object of class "functional_module" and set up various parameters to compute and filter term similarity scores.}
\usage{
merge_pathways(
  object,
  database = NULL,
  p.adjust.cutoff.go = 0.05,
  p.adjust.cutoff.kegg = 0.05,
  p.adjust.cutoff.reactome = 0.05,
  p.adjust.cutoff.hmdb = 0.05,
  p.adjust.cutoff.metkegg = 0.05,
  count.cutoff.go = 5,
  count.cutoff.kegg = 5,
  count.cutoff.reactome = 5,
  count.cutoff.hmdb = 5,
  count.cutoff.metkegg = 5,
  sim.cutoff.go = 0.5,
  sim.cutoff.kegg = 0.5,
  sim.cutoff.reactome = 0.5,
  sim.cutoff.hmdb = 0.5,
  sim.cutoff.metkegg = 0.5,
  measure.method.go = c("Sim_XGraSM_2013", "Sim_Wang_2007", "Sim_Lin_1998",
    "Sim_Resnik_1999", "Sim_FaITH_2010", "Sim_Relevance_2006", "Sim_SimIC_2010",
    "Sim_EISI_2015", "Sim_AIC_2014", "Sim_Zhang_2006", "Sim_universal", "Sim_GOGO_2018",
    "Sim_Rada_1989", "Sim_Resnik_edge_2005", "Sim_Leocock_1998", "Sim_WP_1994",
    "Sim_Slimani_2006", "Sim_Shenoy_2012", "Sim_Pekar_2002", "Sim_Stojanovic_2001",
    "Sim_Wang_edge_2012", "Sim_Zhong_2002", "Sim_AlMubaid_2006", "Sim_Li_2003",
    "Sim_RSS_2013", "Sim_HRSS_2013", "Sim_Shen_2010", "Sim_SSDD_2013", 
    
    "Sim_Jiang_1997", "Sim_Kappa", "Sim_Jaccard", "Sim_Dice", "Sim_Overlap",
    "Sim_Ancestor"),
  control.method.go = list(),
  go.orgdb = NULL,
  measure.method.kegg = c("jaccard", "dice", "overlap", "kappa"),
  measure.method.reactome = c("jaccard", "dice", "overlap", "kappa"),
  measure.method.hmdb = c("jaccard", "dice", "overlap", "kappa"),
  measure.method.metkegg = c("jaccard", "dice", "overlap", "kappa"),
  path = "result",
  save_to_local = FALSE
)
}
\arguments{
\item{object}{An object of class "functional_module", typically a result from enrich_pathway function.}

\item{database}{Character vector, specify which database(s) to use for merging pathways.
\itemize{
\item For genes: 'go', 'kegg', 'reactome'
\item For metabolites: 'hmdb', 'kegg'
}}

\item{p.adjust.cutoff.go}{Adjusted p-value cutoff for GO database. Default is 0.05.}

\item{p.adjust.cutoff.kegg}{Adjusted p-value cutoff for KEGG database. Default is 0.05.}

\item{p.adjust.cutoff.reactome}{Adjusted p-value cutoff for Reactome database. Default is 0.05.}

\item{p.adjust.cutoff.hmdb}{Adjusted p-value cutoff for HMDB pathways from metabolite enrichment analysis result. Default is 0.05.}

\item{p.adjust.cutoff.metkegg}{Adjusted p-value cutoff for KEGG pathways from metabolite enrichment analysis result. Default is 0.05.}

\item{count.cutoff.go}{Gene count cutoff for GO terms from gene enrichment analysis. Default is 5.}

\item{count.cutoff.kegg}{Gene count cutoff for KEGG pathways from gene enrichment analysis. Default is 5.}

\item{count.cutoff.reactome}{Count cutoff for Reactome pathways from gene enrichment analysis. Default is 5.}

\item{count.cutoff.hmdb}{Metabolite count cutoff for HMDB pathways from metabolite enrichment analysis result. Default is 5.}

\item{count.cutoff.metkegg}{Metabolite count cutoff for KEGG pathways from metabolite enrichment analysis result. Default is 5.}

\item{sim.cutoff.go}{Similarity cutoff for GO database. Default is 0.5.}

\item{sim.cutoff.kegg}{Similarity cutoff for KEGG database. Default is 0.5.}

\item{sim.cutoff.reactome}{Similarity cutoff for Reactome database. Default is 0.5.}

\item{sim.cutoff.hmdb}{Similarity cutoff for HMDB database when interpreting metabolite enrichment result. Default is 0.5.}

\item{sim.cutoff.metkegg}{Similarity cutoff for KEGG database when interpreting metabolite enrichment result. Default is 0.5.}

\item{measure.method.go}{A character vector specifying the term semantic similarity measure method for GO terms. Default is \code{"Sim_XGraSM_2013"}. See \code{simona::term_sim()} for available measures.}

\item{control.method.go}{a list of parameters passing to specified measure method for GO term semantic similarity. For details about how to set this parameter, please go to https://jokergoo.github.io/simona/articles/v05_term_similarity.html.}

\item{go.orgdb}{OrgDb object or character string naming the \emph{OrgDb} annotation package used to derive gene–GO mappings, required when database includes "go".}

\item{measure.method.kegg}{A character vector specifying the similarity measure method for KEGG. Choices are "jaccard", "dice", "overlap", "kappa". Default is "jaccard".}

\item{measure.method.reactome}{A character vector specifying the similarity measure method for Reactome. Choices are "jaccard", "dice", "overlap", "kappa". Default is "jaccard".}

\item{measure.method.hmdb}{A character vector specifying the similarity measure method for HMDB when interpreting metabolite enrichment result. Choices are "jaccard", "dice", "overlap", "kappa". Default is "jaccard".}

\item{measure.method.metkegg}{A character vector specifying the similarity measure method for KEGG when interpreting metabolite enrichment result. Choices are "jaccard", "dice", "overlap", "kappa". Default is "jaccard".}

\item{path}{Directory path to save the results. Default is "result".}

\item{save_to_local}{Logical, if TRUE the results will be saved to local disk.}
}
\value{
An object of class "functional_module" with slots for merged pathways from each database.
}
\description{
This function merges enrichment analysis results from different databases (GO, KEGG, and Reactome) into a single object respectively. The function takes an object of class "functional_module" and set up various parameters to compute and filter term similarity scores.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}

Yifei Ge \email{yifeii.ge@outlook.com}
}
