% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_merge_pathways.R
\name{merge_pathways_internal}
\alias{merge_pathways_internal}
\title{Merge Pathway Enrichment Results Internally}
\usage{
merge_pathways_internal(
  pathway_result,
  analysis_type = c("enrich_pathway", "do_gsea"),
  p.adjust.cutoff = 0.05,
  count.cutoff = 5,
  database = c("go", "kegg", "reactome"),
  sim.cutoff = 0.5,
  measure.method,
  control.method.go = list(),
  path = "result",
  save_to_local = FALSE
)
}
\arguments{
\item{pathway_result}{A required object containing results from the \code{enrich_pathway} function.}

\item{analysis_type}{Character, type of analysis to perform: either \code{"enrich_pathway"} or \code{"do_gsea"}.}

\item{p.adjust.cutoff}{Numeric, p-adjusted value cutoff for filtering enriched pathways.}

\item{count.cutoff}{Numeric, count cutoff for filtering enriched pathways.}

\item{database}{Character vector, the database from which the enrichment results were obtained ('go', 'kegg', 'reactome').}

\item{sim.cutoff}{Numeric, similarity cutoff for clustering pathways.}

\item{measure.method}{Character, method for calculating term similarity.
Default is \code{"Sim_XGraSM_2013"}. See \code{simona::all_term_sim_methods()} for available measures.}

\item{control.method.go}{a list of parameters passing to specified measure method for GO term semantic similarity. For details about how to set this parameter, please go to https://jokergoo.github.io/simona/articles/v05_term_similarity.html.}

\item{path}{Character, directory to save intermediate and final results.}

\item{save_to_local}{Logical, if TRUE the results will be saved to local disk.}
}
\value{
A list containing \code{graph_data}, \code{module_result}, and \code{result_with_module}.
}
\description{
This function merges pathway enrichment results obtained through various databases (GO, KEGG, Reactome).
It applies similarity measures to find closely related pathways and categorizes them into modules.
}
\examples{
\dontrun{
# Load pathway results obtained through `enrich_pathway` function
pathway_results <- load_pathway_results("path/to/results")

# Merge pathways and find modules
merged_results <- merge_pathways_internal(pathway_result = pathway_results)
}
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
