% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_llm_module_output_generation.R
\name{module_name_generation}
\alias{module_name_generation}
\title{Module Name Generation}
\usage{
module_name_generation(paper_result, api_key)
}
\arguments{
\item{paper_result}{A list containing module information, including related papers and PubMed results.}

\item{api_key}{A string containing the API key required to access the AI API.}
}
\value{
A list of results for each module, where each element is a list containing
\code{module_name} and \code{summary}.
}
\description{
This internal function processes a list of modules to generate final biological module names
and summaries for all modules by calling \code{single_module_generation}.
}
\seealso{
\code{\link{single_module_generation}}
}
\author{
Feifan Zhang \email{FEIFAN004@e.ntu.edu.sg}
}
\keyword{internal}
