% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-data_visualization_plot_module_info.R
\name{plot_module_info}
\alias{plot_module_info}
\title{Plot Module Information for Enrichment Analysis}
\usage{
plot_module_info(
  object,
  level = c("module", "functional_module"),
  database = c("go", "kegg", "reactome"),
  module_id,
  translation = FALSE
)
}
\arguments{
\item{object}{An object of a specific class containing enrichment results.}

\item{level}{A character string specifying the level of biological organization, either "module" or "functional_module".}

\item{database}{A character string specifying the source database, either "go", "kegg" or "reactome".}

\item{module_id}{A single identifier specifying the module of interest.}

\item{translation}{translation or not.}
}
\value{
A list containing three ggplot objects: network plot (\code{network}), bar plot (\code{barplot}), and word cloud (\code{wordcloud}).
}
\description{
This function generates various visualizations including a network plot, bar plot,
and word cloud to provide detailed insights into the enriched functional or biological modules.
}
\examples{
\dontrun{
result <- plot_module_info(object = myObject,
level = "module",
database = "go",
module_id = "M123")
}
}
\author{
Xiaotao Shen \href{mailto:shenxt1990@outlook.com}{shenxt1990@outlook.com}
}
