% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-data_visualization_plot_pathway_bar.R
\name{plot_pathway_bar}
\alias{plot_pathway_bar}
\title{Plot Pathway Bar Chart}
\usage{
plot_pathway_bar(
  object,
  top_n = 10,
  y_lable_width = 50,
  translation = FALSE,
  translate_to = c("chinese", "spanish", "english", "french", "german", "italian",
    "japanese", "korean", "portuguese", "portuguese", "russian", "spanish"),
  engine = c("gemini", "chatgpt"),
  level = c("pathway", "module", "functional_module"),
  line_type = c("straight", "meteor"),
  p.adjust.cutoff = 0.05,
  count.cutoff = 5,
  database_color = c(GO = "#1F77B4FF", KEGG = "#FF7F0EFF", Reactome = "#2CA02CFF"),
  database = c("go", "kegg", "reactome")
)
}
\arguments{
\item{object}{An object containing the enrichment results and other relevant data.}

\item{top_n}{An integer specifying the top N pathways to display.}

\item{y_lable_width}{An integer specifying the width of the Y-axis labels.}

\item{translation}{translation or not.}

\item{translate_to}{translate to which language.}

\item{engine}{translate engine.}

\item{level}{A character string specifying the level of analysis.
One of "pathway", "module", or "functional_module".}

\item{line_type}{A character string specifying the type of line to use for the bar chart.
One of "straight" or "meteor".}

\item{p.adjust.cutoff}{A numeric value for the FDR adjusted P-value cutoff.}

\item{count.cutoff}{A numeric value for the count cutoff.}

\item{database_color}{A named vector containing the colors for different databases.}

\item{database}{A character vector specifying which databases to include in the plot.
Should be "go", "kegg", or/and "reactome".}
}
\value{
A ggplot object representing the bar chart.
}
\description{
This function plots a bar chart representing enriched pathways, modules, or functional modules.
The bars are colored according to the database of origin.
}
\examples{
data("enriched_functional_module", package = "mapa")
plot_pathway_bar(
  object = enriched_functional_module,
  top_n = 10,
  y_lable_width = 30,
  level = "pathway"
)

plot_pathway_bar(
  object = enriched_functional_module,
  top_n = 10,
  y_lable_width = 30,
  level = "pathway",
  line_type = "meteor"
)

plot_pathway_bar(enriched_functional_module,
                 top_n = 10,
                 level = "module")

plot_pathway_bar(
  enriched_functional_module,
  top_n = 10,
  level = "functional_module",
  p.adjust.cutoff = 0.05
)
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
