% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/15-data_visualization_plot_relationship_network.R
\name{plot_relationship_network}
\alias{plot_relationship_network}
\title{Plot Relationship Network}
\usage{
plot_relationship_network(
  object,
  include_functional_modules = TRUE,
  include_modules = TRUE,
  include_pathways = TRUE,
  include_molecules = TRUE,
  functional_module_color = "#F05C3BFF",
  module_color = "#46732EFF",
  pathway_color = "#197EC0FF",
  molecule_color = "#3B4992FF",
  functional_module_text = TRUE,
  module_text = TRUE,
  pathway_text = TRUE,
  molecule_text = FALSE,
  functional_module_text_size = 3,
  module_text_size = 3,
  pathway_text_size = 3,
  molecule_text_size = 3,
  circular_plot = FALSE,
  functional_module_arrange_position = TRUE,
  module_arrange_position = TRUE,
  pathway_arrange_position = TRUE,
  molecule_arrange_position = TRUE,
  functional_module_position_limits = c(0, 1),
  module_position_limits = c(0, 1),
  pathway_position_limits = c(0, 1),
  molecule_position_limits = c(0, 1)
)
}
\arguments{
\item{object}{An object of class "functional_module".}

\item{include_functional_modules}{Logical; include functional modules in the plot?}

\item{include_modules}{Logical; include modules in the plot?}

\item{include_pathways}{Logical; include pathways in the plot?}

\item{include_molecules}{Logical; include molecules in the plot?}

\item{functional_module_color}{Character; color for functional modules.}

\item{module_color}{Character; color for modules.}

\item{pathway_color}{Character; color for pathways.}

\item{molecule_color}{Character; color for molecules.}

\item{functional_module_text}{Logical; include functional module text labels?}

\item{module_text}{Logical; include module text labels?}

\item{pathway_text}{Logical; include pathway text labels?}

\item{molecule_text}{Logical; include molecule text labels?}

\item{functional_module_text_size}{Numeric; size of functional module text labels.}

\item{module_text_size}{Numeric; size of module text labels.}

\item{pathway_text_size}{Numeric; size of pathway text labels.}

\item{molecule_text_size}{Numeric; size of molecule text labels.}

\item{circular_plot}{Logical; make the plot circular?}

\item{functional_module_arrange_position}{Logical; should the position of functional modules be arranged?}

\item{module_arrange_position}{Logical; should the position of modules be arranged?}

\item{pathway_arrange_position}{Logical; should the position of pathways be arranged?}

\item{molecule_arrange_position}{Logical; should the position of molecules be arranged?}

\item{functional_module_position_limits}{Numeric vector; limits for functional module positions.}

\item{module_position_limits}{Numeric vector; limits for module positions.}

\item{pathway_position_limits}{Numeric vector; limits for pathway positions.}

\item{molecule_position_limits}{Numeric vector; limits for molecule positions.}
}
\value{
A ggplot object representing the relationship network.
}
\description{
This function plots a relationship network that includes functional modules,
modules, pathways, and molecules.
}
\author{
Xiaotao Shen \email{shenxt1990@outlook.com}
}
