% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_llm_module_online_retrieval.R
\name{process_module}
\alias{process_module}
\title{Process a Single Module to Retrieve PubMed IDs}
\usage{
process_module(module_name, module, chunk_size = 5, years = 5, retmax = 10)
}
\arguments{
\item{module_name}{A character string specifying the name of the module.}

\item{module}{A list containing module information:
\itemize{
\item For gene modules (length 6): Contains PathwayNames, GeneSymbols, and GeneNames_vec
\item For metabolite modules (length 5): Contains PathwayNames and MetNames_vec
}}

\item{chunk_size}{An integer specifying the size of query chunks (default is 5).}

\item{years}{An integer specifying how many years to look back in the search (default is 5).}

\item{retmax}{An integer specifying the maximum number of results to retrieve (default is 10).}
}
\value{
A list containing:
\item{module_name}{The name of the module.}
\item{PubmedIDs}{A character vector of unique PubMed IDs retrieved for the module.}
}
\description{
This internal function processes a module to query PubMed IDs based on either gene information
or metabolite information, combined with pathway data. It detects the module type automatically
based on the module's structure.
}
\author{
Feifan Zhang \email{FEIFAN004@e.ntu.edu.sg}

Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
