% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_llm_module_get_pathway_name_desc_pmid.R
\name{quickgo_api}
\alias{quickgo_api}
\title{Query the QuickGO API for GO term information}
\usage{
quickgo_api(go_ids)
}
\arguments{
\item{go_ids}{A character vector of GO IDs}
}
\value{
A list containing the query results from the QuickGO API
}
\description{
This internal function sends a request to the QuickGO API and retrieves
information about GO terms.
}
\author{
Yifei Ge \email{yifeii.ge@outlook.com}
}
\keyword{internal}
