% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_mnetwork_class.R
\name{slot-replace}
\alias{slot-replace}
\alias{mol_nodes<-}
\alias{mol_layers<-}
\alias{mol_layers_edge_weight<-}
\alias{path_nodes<-}
\alias{path_layer_edge_weight<-}
\alias{path_to_mol<-}
\alias{pathway_mol_edge_weight<-}
\alias{layer_weights<-}
\alias{mol_nodes<-,multi_omics_functional_module-method}
\alias{mol_layers<-,multi_omics_functional_module-method}
\alias{path_nodes<-,multi_omics_functional_module-method}
\alias{path_layer_edge_weight<-,multi_omics_functional_module-method}
\alias{path_to_mol<-,multi_omics_functional_module-method}
\alias{pathway_mol_edge_weight<-,multi_omics_functional_module-method}
\alias{layer_weights<-,multi_omics_functional_module-method}
\title{Replace slots of a multi_omics_functional_module object}
\usage{
mol_nodes(x) <- value

mol_layers(x) <- value

mol_layers_edge_weight(x) <- value

path_nodes(x) <- value

path_layer_edge_weight(x) <- value

path_to_mol(x) <- value

pathway_mol_edge_weight(x) <- value

layer_weights(x) <- value

\S4method{mol_nodes}{multi_omics_functional_module}(x) <- value

\S4method{mol_layers}{multi_omics_functional_module}(x) <- value

\S4method{path_nodes}{multi_omics_functional_module}(x) <- value

\S4method{path_layer_edge_weight}{multi_omics_functional_module}(x) <- value

\S4method{path_to_mol}{multi_omics_functional_module}(x) <- value

\S4method{pathway_mol_edge_weight}{multi_omics_functional_module}(x) <- value

\S4method{layer_weights}{multi_omics_functional_module}(x) <- value
}
\arguments{
\item{x}{A \code{multi_omics_functional_module} object.}

\item{value}{New value; must match the type of the corresponding slot.}
}
\value{
The updated \code{multi_omics_functional_module} object.
}
\description{
Each replacement method validates the updated object via \code{\link[=validObject]{validObject()}}
before returning it.
}
