% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100_translation.R
\name{translate_language}
\alias{translate_language}
\alias{translate_language.character}
\title{Translate Language}
\usage{
translate_language(
  text,
  engine = c("gemini", "chatgpt"),
  to = c("chinese", "spanish", "english", "french", "german", "italian", "japanese",
    "korean", "portuguese", "portuguese", "russian", "spanish")
)

\method{translate_language}{character}(
  text,
  engine = c("gemini", "chatgpt"),
  to = c("chinese", "spanish", "english", "french", "german", "italian", "japanese",
    "korean", "portuguese", "portuguese", "russian", "spanish")
)
}
\arguments{
\item{text}{Text to be translated.}

\item{engine}{Translation engine to use, defaults to 'gemini' or 'chatgpt'.}

\item{to}{Target language for translation, options include 'chinese', 'spanish', 'english',
'french', 'german', 'italian', 'japanese', 'korean', 'portuguese', 'russian', and 'spanish'.}
}
\value{
An object of the same class as \code{text} with the translated content.
}
\description{
This function serves as a generic method for translating text using different engines.
It requires a specific method to be defined for each type of input (e.g., character, functional_module).
}
\examples{
\dontrun{
  # Translate text to French using Gemini
  translated_text <- translate_language(text = "Hello World", engine = "gemini", to = "french")
}
}
