% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100-translation.R
\name{translate_language}
\alias{translate_language}
\title{Translate Text into Different Languages}
\usage{
translate_language(
  text =
    "As of my last update in early 2023, there's no widely\\n           recognized individual named Xiaotao Shen affiliated with Stanford\\n           University that I can refer to.",
  engine = c("gemini", "chatgpt"),
  to = c("chinese", "spanish", "english", "french", "german", "italian", "japanese",
    "korean", "portuguese", "portuguese", "russian", "spanish")
)
}
\arguments{
\item{text}{A string representing the text to be translated.
Default is a description of the Gemini model.}

\item{engine}{A string specifying the translation engine to be used, "chatgpt" or "gemini".}

\item{to}{A string indicating the target language for translation.
Supported languages include 'chinese', 'spanish', 'english', 'french', 'german',
'italian', 'japanese', 'korean', 'portuguese', 'russian', and 'spanish'.}
}
\value{
A character vector containing the translated text.
}
\description{
This function translates a given text into a specified language using the Gemini translation engine.
It utilizes the Gemini large language models for translation.
}
\examples{
\dontrun{
  # Translate text to French using Gemini
  translated_text <- translate_language(text = "Hello World", engine = "gemini", to = "french")
}

}
