% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-google_gemini.R
\name{translate_language}
\alias{translate_language}
\title{Translate Text into Different Languages}
\usage{
translate_language(
  text =
    "Gemini is a family of multimodal large language models\\n           developed by Google DeepMind,\\n           serving as the successor to LaMDA and PaLM 2.\\n           Comprising Gemini Ultra, Gemini Pro, and Gemini Nano,\\n           it was announced on December 6, 2023,\\n           positioned as a contender to OpenAI's GPT-4.",
  engine = c("gemini"),
  to = c("chinese", "spanish", "english", "french", "german", "italian", "japanese",
    "korean", "portuguese", "russian", "spanish")
)
}
\arguments{
\item{text}{A string representing the text to be translated.
Default is a description of the Gemini model.}

\item{engine}{A string specifying the translation engine to be used.
Currently, only 'gemini' is supported.}

\item{to}{A string indicating the target language for translation.
Supported languages include 'chinese', 'spanish', 'english', 'french', 'german',
'italian', 'japanese', 'korean', 'portuguese', 'russian', and 'spanish'.}
}
\value{
A character vector containing the translated text.
}
\description{
This function translates a given text into a specified language using the Gemini translation engine.
It utilizes the Gemini large language models for translation.
}
\examples{
\dontrun{
  # Translate text to French using Gemini
  translated_text <- translate_language(text = "Hello World", engine = "gemini", to = "french")
}

}
